[merge] merge with 정용, add spell_checker directory for clean, refactor locations
